Use add methods of JavaClass to generate all classes
Milestone in the usage of JavaClass instead of representing classes with a string only.
Some of the old code is still there so comparisons between the new and the old way is easily done.
Some old methods will need to be changed to fit better with the new design,
such as generate_child which should modify a JavaClass object instead of building an
"access_methods" string