Add gen data attribute to MethodGenerator class
This attribute is supposed to contain an instance of the appropriate subclass for
instances of the base class, and be equal to self for the others. There is a risk
of infinite recursion doing this, so assertions should be added once there are
method generator subclasses for containers and lists as well.