Add support for importing other modules
This should have been done a long time ago...
The module name is now used as subpackage of the user provided package name
This means that classes can be generated for multiple modules
This fix breaks some previous code:
- package-info.java file generation won't work for imported and augmented module package hierarchies
- build files and launch configurations for examples and tests will have to be updated
- built in javadoc generation probably won't work (but is not needed anyway, so it will be removed shortly)

The change also include some unrelated bug fixes:
- Naming the root class after the prefix and not the module name
(which was temporarily assumed to be the same for some strange reason), and augmentation related stuff.

There might be some other changes not mentioned here as well, since
the change is so big it was hard to divide it into smaller commits.