Fix issue with unknown namespace for submodules
Use module name to search in ctx dict for the module containing the namespace
Beware that the revision of the module is ignored, and that the search is performed in time linear to the number of modules.
It might be a good idea to cache the parent module in the submodule statement.