Convert identifiers to avoid name clashes
If identifiers of module, prefix, containers, etc. begins with
an upper case letter, the generated package names can clash
with class names. This commit provides a work-around for this issue.

Note that no caching is used and the entire tree of statements
is traversed for every included module, so this could be a
performance issue if the model is large